+++ documentation/demopages/experimenteel/algemeen/Loading.tsx
@@ -14,18 +14,37 @@ const Loading = () => {
         <LayoutFlow>
           <div className="rvo-loading-canvas rvo-loading-canvas-animate">
             <Icon icon="pijlen-in-cirkel-om-document" size="2xl" color="grijs-700"></Icon>
-            <p className="rvo-loading-text rvo-paragraph rvo-paragraph--grijs-500 rvo-paragraph--sm rvo-paragraph--no-spacing">
+            <p className="rvo-loading-text rvo-paragraph rvo-paragraph--sm rvo-margin-block-start--sm">
               De bestanden worden geladen...
             </p>
           </div>
-          <div className="rvo-loading-canvas">
+          <div className="rvo-loading-canvas rvo-loading-canvas-animate">
             <Icon icon="pijlen-in-cirkel-om-document" size="2xl" color="grijs-700"></Icon>
-            <p className="rvo-loading-text rvo-paragraph rvo-paragraph--grijs-500 rvo-paragraph--sm rvo-paragraph--no-spacing">
+            <p className="rvo-loading-text rvo-paragraph rvo-paragraph--sm rvo-margin-block-end--sm rvo-margin-block-start--sm">
               <LayoutFlow row={true} gap="xs">
-                <StatusIcon type="waarschuwing" size="sm"></StatusIcon>Helaas konden de bestanden niet geladen worden.
+                <StatusIcon type="waarschuwing" size="sm"></StatusIcon>Het laden duurt langer dan verwacht...
               </LayoutFlow>
             </p>
-            <Button size="xs">Probeer het opnieuw</Button>
+            <LayoutFlow row={true} gap="md">
+              <Button size="xs">Probeer het opnieuw</Button>
+              <Button size="xs" kind="secondary">
+                Probleem melden
+              </Button>
+            </LayoutFlow>
+          </div>
+          <div className="rvo-loading-canvas">
+            <Icon icon="pijlen-in-cirkel-om-document" size="2xl" color="grijs-700"></Icon>
+            <p className="rvo-loading-text rvo-paragraph rvo-paragraph--sm rvo-margin-block-end--sm rvo-margin-block-start--sm">
+              <LayoutFlow row={true} gap="xs">
+                <StatusIcon type="foutmelding" size="sm"></StatusIcon>Helaas konden de bestanden niet geladen worden.
+              </LayoutFlow>
+            </p>
+            <LayoutFlow row={true} gap="md">
+              <Button size="xs">Probeer het opnieuw</Button>
+              <Button size="xs" kind="secondary">
+                Probleem melden
+              </Button>
+            </LayoutFlow>
           </div>
         </LayoutFlow>
       </main>