+++ documentation/demopages/voorbeelden/cmor/DocumentenClean.tsx
@@ -107,7 +107,10 @@ const DocumentenClean = () => {
               <div>
                 <LayoutFlow gap="xl">
                   <Tabs
-                    tabs={[{ label: 'Mijn ontvangen documenten' }, { label: 'Mijn verstuurde documenten' }]}
+                    tabs={[
+                      { label: 'Mijn ontvangen documenten', href: '#' },
+                      { label: 'Mijn verstuurde documenten', href: '#' },
+                    ]}
                     activeTab={0}
                   />
                   <Grid columns="four">
@@ -327,45 +330,7 @@ const DocumentenClean = () => {
           </div>
         </div>
       </main>
-      <Footer
-        columns={[
-          {
-            label: 'RVO',
-            items: [
-              { content: 'Over ons', link: '#' },
-              { content: 'Contact', link: '#' },
-              { content: 'Pers', link: '#' },
-              { content: 'Zaken regelen bij RVO', link: '#' },
-            ],
-          },
-          {
-            label: 'Over deze site',
-            items: [
-              { content: 'Privacy', link: '#' },
-              { content: 'Cookies en anti-spam', link: '#' },
-              { content: 'Toegankelijkheid', link: '#' },
-              { content: 'Proclaimer', link: '#' },
-            ],
-          },
-          {
-            label: 'Snel naar',
-            items: [
-              { content: 'Onderwerpen A t/m Z', link: '#' },
-              { content: 'Klantenpanel', link: '#' },
-              { content: 'Documenten en publicaties', link: '#' },
-              { content: 'Overheidspublicaties en auteursrecht', link: '#' },
-            ],
-          },
-          {
-            label: 'Blijf op de hoogte',
-            items: [
-              { content: 'Social media', link: '#' },
-              { content: 'RSS', link: '#' },
-              { content: 'Gratis nieuwsbrief', link: '#' },
-            ],
-          },
-        ]}
-      />
+      <Footer />
     </body>
   );
 };